pyproject.toml
[tool.poetry]
name = "mbmp"
version = "0.1.0"
description = "Exercises of Model Building in Mathematical Programming."
authors = ["sfujiwara <shuhei.fujiwara@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pulp = "^2.1"
pyyaml = "^5.3.1"
protobuf = "^3.11.3"
jupyter = "^1.0.0"
numpy = "^1.18.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
